pyproject.toml
[project]
name = "pulso-electoral"
version = "0.1.0"
description = "Social Listening & Digital Manipulation Research for Colombia's 2026 Elections — CIVICUS DDI"
authors = [
    "MGO <mgo@pulsoelectoral.org>",
]
requires-python = ">=3.12"
readme = "README.md"

dependencies = [
    "numpy>=2.4.2",
    "pandas[parquet]>=2.3.0",
    "feedparser>=6.0.11",
    "praw>=7.8.1",
    "requests>=2.32.0",
    "duckdb>=1.2.0",
    "pysentimiento>=0.7.0",
    "spacy>=3.8.0",
    "sentence-transformers>=3.4.0",
    "langdetect>=1.0.9",
    "pyyaml>=6.0.2",
    "python-dotenv>=1.0.1",
]

[project.urls]
repository = "https://github.com/pulso-electoral/pulso-electoral"

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "pytest-asyncio>=0.25.0",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.0",
    "seaborn>=0.13.2",
    "plotly>=6.1.0",
    "ruff>=0.12.0",
    "mypy>=1.17.0",
    "types-requests>=2.32.0",
]

docs = [
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.7.1",
    "mkdocstrings[python]>=1.0.3",
]

app = [
    "streamlit>=1.45.0",
    "plotly>=6.1.0",
    "folium>=0.19.0",
    "streamlit-folium>=0.24.0",
]

[tool.uv]
package = false
default-groups = [
    "dev",
    "docs",
    "app",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.coverage.paths]
source = ["src/"]

[tool.coverage.run]
source = ["src"]
omit = ["**/__init__.py"]
